I sleeeeeep

+++ src/main/java/com/t/library/Utils/HTTPUtils.java
@@ -27,7 +27,7 @@ public class HTTPUtils {
 
         }
     }
-    public String delete(String url, Long id) throws IOException {
+    public String delete(String url, int id) throws IOException {
         Request req = new Request.Builder().url(url + id).delete().build();
         try (Response response = client.newCall(req).execute()) {
             return response.body().string();
